global: configure and fixes gometalinter issues
