+++ .gometalinter.json
@@ -1,0 +1,4 @@
+{
+    "Disable": ["gocyclo", "maligned"],
+    "Skip": ["vendor"]
+}
